+++ 3esCore/3esRuntime/Handlers/Shape3D/StarHandler.cs
@@ -36,9 +36,10 @@ namespace Tes.Handlers.Shape3D
     {
       transform = Matrix4x4.identity;
 
-      float radius = attributes.ScaleX;
-      transform.SetColumn(3, new Vector4(attributes.X, attributes.Y, attributes.Z, 1.0f));
-      var pureRotation = Matrix4x4.Rotate(new Quaternion(attributes.RotationX, attributes.RotationY, attributes.RotationZ, attributes.RotationW));
+      float radius = (float)attributes.ScaleX;
+      transform.SetColumn(3, new Vector4((float)attributes.X, (float)attributes.Y, (float)attributes.Z, 1.0f));
+      var pureRotation = Matrix4x4.Rotate(new Quaternion((float)attributes.RotationX, (float)attributes.RotationY,
+                                                         (float)attributes.RotationZ, (float)attributes.RotationW));
       transform.SetColumn(0, pureRotation.GetColumn(0) * radius);
       transform.SetColumn(1, pureRotation.GetColumn(1) * radius);
       transform.SetColumn(2, pureRotation.GetColumn(2) * radius);
